+++ Packages/com.unity.inputsystem/InputSystem/Actions/Interactions/HoldInteraction.cs
@@ -1,4 +1,3 @@
-using System;
 using System.ComponentModel;
 using UnityEngine.Scripting;
 #if UNITY_EDITOR